Laser improvements and other things.
Changed laser material to a translusive one, and odded a fresnel effect to it.
Added a lightsource to laser blueprint to counteracct that it no longer gave light to scene.
Changed door open/close timings.
Disabled autoexposure in project settings.
Added a TestMap for logic componenets.
Disabled collision of pushablebox boxcollider with quantum light trace.